+++ src/composables/useTrackers.ts
@@ -1,9 +1,10 @@
 import { computed, ref, watch, type Ref } from 'vue';
 
 import { getTrackersForFlaws, type TrackersPost, fileTrackingFor } from '@/services/TrackerService';
-import type { ZodAffectType } from '@/types/zodAffect';
+import type { ZodAffectType, ZodTrackerType } from '@/types/zodAffect';
 
 export type UpdateStream = ModuleComponentProductStream & UpdateStreamMeta;
+type ZodTrackerTypeWithAffect = ZodTrackerType & ZodAffectType;
 
 type ModuleComponent = {
   ps_module: string;
@@ -22,8 +23,9 @@ type ModuleComponentProductStream = {
 };
 
 type UpdateStreamMeta = {
-  affectUuid?: string;
+  affectUuid?: string | null;
   ps_component?: string;
+  ps_module?: string;
 };
 
 export function useTrackers(flawUuid: string, affects: Ref<ZodAffectType[]>) {
@@ -51,25 +53,22 @@ export function useTrackers(flawUuid: string, affects: Ref<ZodAffectType[]>) {
     })
   );
 
-  const availableUpdateStreams = computed((): UpdateStream[] => moduleComponents.value.flatMap(
-    (moduleComponent: any) =>
-      moduleComponent.streams.map((stream: Record<string, any>) => (
-        {
-          ...stream,
-          ps_component: moduleComponent.ps_component,
-          affectUuid: moduleComponent.affect.uuid
-        }
-      ))
-  )
-    .filter((stream: any) => !alreadyFiledTrackers.value.find(
-      (tracker: any) => tracker.ps_update_stream === stream.ps_update_stream
-        && tracker.ps_component === stream.ps_component
-    ))
-  );
+  const availableUpdateStreams = computed((): UpdateStream[] => moduleComponents.value.flatMap((moduleComponent) =>
+    moduleComponent.streams.map((stream: UpdateStream) => ({
+      ...stream,
+      ps_component: moduleComponent.ps_component,
+      ps_module: moduleComponent.ps_module,
+      affectUuid: moduleComponent.affect.uuid
+    }))
+  ).filter((stream: UpdateStream) => !alreadyFiledTrackers.value.find(
+    (tracker: ZodTrackerTypeWithAffect) => tracker.ps_update_stream === stream.ps_update_stream
+      && tracker.ps_component === stream.ps_component
+  )));
 
   const sortedStreams = computed(
     (): UpdateStream[] => availableUpdateStreams.value
-      .toSorted((a, b) => a.ps_update_stream.localeCompare(b.ps_update_stream))
+      .slice()
+      .sort((a, b) => a.ps_update_stream.localeCompare(b.ps_update_stream))
       .filter((tracker) =>
         filterString.value === ''
           ? tracker
@@ -87,6 +86,16 @@ export function useTrackers(flawUuid: string, affects: Ref<ZodAffectType[]>) {
   );
 
   const filterString = ref('');
+
+  const untrackedAffects = computed(() => affects.value.filter((affect) => affect.trackers.length === 0));
+
+  const untrackableAffects = computed(() => untrackedAffects.value
+    .filter(
+      (affect) => availableUpdateStreams.value.find(
+        stream => stream.ps_module === affect.ps_module && stream.ps_component === affect.ps_component
+      ) === undefined
+    )
+  );
 
   watch(moduleComponents, () => {
     moduleComponents.value.forEach((moduleComponent) => {
@@ -131,7 +140,7 @@ export function useTrackers(flawUuid: string, affects: Ref<ZodAffectType[]>) {
 
   function getUpdateStreamsFor(module: string, component: string) {
     const moduleComponent = moduleComponents.value.find(
-      ({ ps_module, ps_component }: any) => ps_module === module && ps_component === component
+      ({ ps_module, ps_component }: ModuleComponent) => ps_module === module && ps_component === component
     );
     return moduleComponent ? moduleComponent.streams : [];
   }
@@ -148,6 +157,7 @@ export function useTrackers(flawUuid: string, affects: Ref<ZodAffectType[]>) {
     getUpdateStreamsFor,
     alreadyFiledTrackers,
     availableUpdateStreams,
+    untrackableAffects,
     trackerSelections,
     trackersToFile,
     setAllTrackerSelections,